+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -  Basic application structure for interacting with the cli
 -  Open & Parse Schema
+-  Add a Cargo.toml generation
 
 ## [0.1.0]